requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alabaster==1.0.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
azure-ai-inference==1.0.0b8
azure-core==1.32.0
babel==2.17.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
blessed==1.20.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
debugpy==1.8.12
deepsearch-glm==1.0.0
defusedxml==0.7.1
dill==0.3.9
distro==1.9.0
dj-rest-auth==6.0.0
Django==4.2.7
django-allauth==0.53.0
django-js-asset==2.2.0
django-markdownx==4.0.7
django-picklefield==3.2
django-prometheus==2.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
docling==2.17.0
docling-core==2.16.1
docling-ibm-models==3.3.0
docling-parse==3.1.2
docutils==0.21.2
easyocr==1.7.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-documentai==2.32.0
google-genai==1.3.0
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.26.0
idna==3.7
imageio==2.37.0
imagesize==1.4.1
isodate==0.7.2
Jinja2==3.1.5
jiter==0.5.0
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lazy_loader==0.4
llamaapi==0.1.36
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownify==1.0.0
marko==2.1.2
MarkupSafe==3.0.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.65.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.1.2
pypdfium2==4.30.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
safetensors==0.5.2
scikit-image==0.25.1
scipy==1.15.1
semchunk==2.2.2
shapely==2.0.7
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-autobuild==2024.10.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
tifffile==2025.1.10
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.5
transformers==4.48.2
triton==3.2.0
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
XlsxWriter==3.2.2
yarl==1.18.3
